backend/requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
groq==0.4.2
supabase==2.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
